backend/document_translate/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
torch
transformers
sentencepiece
pdfplumber
python-docx
